Fix test test_template.py::test_template (#19)
* Remove obsolete version test

Deleted the test_version.py file as it was checking for a fixed version string of "unknown," which is no longer relevant. This change helps clean up the test suite by removing unnecessary tests that do not contribute to meaningful validation.

* Add detailed assert error messages in tests

Enhanced test assertions with specific error messages to clarify differences in test outcomes. These changes improve debugging by specifying which parameters or options differ between expected and actual results. This will aid in quickly identifying and resolving issues during test failures.

* Remove redundant checklist item in PR template

The checklist item for adding code comments to explain the intent was duplicated. This change removes the redundant item to improve clarity and maintainability of the pull request template. All contributors are encouraged to check the remaining items before submitting a PR.

* Update Conda setup in GitHub Actions workflow

Replace `mamba-version` with `miniforge-version` for improved dependency management. Remove caching keys for environment and downloads, streamlining the workflow configuration. This change aims to ensure compatibility with the latest tools and simplify maintenance.

* ensure that the script is cleaned from the list of modules

Signed-off-by: Jose Borreguero <borreguero@gmail.com>